feat: move fillBrush into Annotation Class
